Implementing moduler backend design.
