+++ assets/buttons.js
@@ -10,12 +10,12 @@ function openPage(pageName, elmnt, color) {
     for (i = 0; i < tablinks.length; i++) {
       tablinks[i].style.backgroundColor = "";
     }
-    document.getElementById(pageName).style.display = "block";
-    elmnt.style.backgroundColor = color;
+    
+    if (!(pageName === undefined)) {
+      document.getElementById(pageName).style.display = "block";
+      elmnt.style.backgroundColor = color;
+    }
   }
-  $("#searchBtn").on("click", function (event) {
-    event.preventDefault();
-    getLyrics();
-  });
+  
   // Hides everything when yoou click the tabs
-  openPage()
+  openPage();
